sdp_project_demo now uses KinematicCarTwoPlanners
KinematicCarTwoPlanners has functions to support multiple planners. 

Tested in gui, created odd plan for moving from 0,0,0 to 0,0,0 in png.

__init__.py includes KinematicCarTwoPlanners.

Time to connect the gui functions working on #11 

Closes #12 :
Variable planners are not passed as parameters but set function 
for each planner allows different planners to be set.

For more on the multiple planners see commits:
29cc8659882d13c70734b9a78770e0fd68763b11 and
12868a247f91727e7659c927dc7729a7b48c7ffa
as well as issues #26 and #23 also related to multiple planners.